% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\docType{package}
\name{MBbioeq-package}
\alias{MBbioeq}
\alias{MBbioeq-package}
\title{MBbioeq: Model-Based Bioequivalence Analysis Using SAEM and TOST}
\description{
Performs model-based bioequivalence analysis for parallel-arm studies using Stochastic Approximation Expectation-Maximization (SAEM) estimation and Two One-Sided Tests (TOST) procedure. Supports 1- and 2-compartment pharmacokinetic models with optional absorption lag time. Implements Normal and Gallant corrections for TOST, and delta-method for Cmax variance calculation. Allows flexible random effects structures (diagonal, full, block) and multiple error models (additive, proportional, combined).
}
\author{
\strong{Maintainer}: Eden BETE \email{eden.bete.etu@univ-lille.fr}

Authors:
\itemize{
  \item Celia BENGUENNA \email{celia.benguenna.etu@univ-lille.fr}
}

}
\keyword{internal}
